backend/requirements.txt
streamlit
email-validator
cryptography
python-dotenv
flask
flask-cors
supabase
python-dotenv
torch
requests